main: cleanup options.h header inclusions
To simplify the interface between main part and parsers,
this change cleans up optoins.h.

All main private things are moved to optoins_p.h header.
