+++ main/sort.c
@@ -26,7 +26,7 @@
 
 #include "debug.h"
 #include "entry.h"
-#include "options.h"
+#include "options_p.h"
 #include "read.h"
 #include "routines.h"
 #include "sort.h"
