+++ main/writer-ctags.c
@@ -11,7 +11,7 @@
 
 #include "entry.h"
 #include "mio.h"
-#include "options.h"
+#include "options_p.h"
 #include "read.h"
 #include "ptag.h"
 #include "writer.h"
