+++ main/writer-etags.c
@@ -14,7 +14,7 @@
 #include "debug.h"
 #include "entry.h"
 #include "mio.h"
-#include "options.h"
+#include "options_p.h"
 #include "read.h"
 #include "vstring.h"
 #include "writer.h"
