+++ win32/ctags_vs2013.vcxproj
@@ -275,6 +275,7 @@
     <ClInclude Include="..\main\numarray.h" />
     <ClInclude Include="..\main\objpool.h" />
     <ClInclude Include="..\main\options.h" />
+    <ClInclude Include="..\main\options_p.h" />
     <ClInclude Include="..\main\param.h" />
     <ClInclude Include="..\main\parse.h" />
     <ClInclude Include="..\main\parsers.h" />
